add missing_policy.py
