+++ stix2elevator/convert_stix.py
@@ -54,13 +54,19 @@ from stix2elevator.ids import (add_id_value,
                                get_id_values,
                                get_type_from_id,
                                record_ids)
+from stix2elevator.missing_policy import (convert_to_custom_property_name,
+                                          handle_missing_confidence_property,
+                                          handle_missing_statement_properties,
+                                          handle_missing_string_property,
+                                          handle_missing_tool_property,
+                                          handle_multiple_missing_statement_properties,)
 from stix2elevator.options import error, get_option_value, info, warn
 from stix2elevator.utils import (add_marking_map_entry,
                                  check_map_1x_markings_to_2x,
                                  convert_controlled_vocabs_to_open_vocabs,
                                  convert_timestamp_of_stix_object,
                                  convert_timestamp_to_string,
-                                 convert_to_custom_property_name,
+
                                  identifying_info,
                                  iterpath,
                                  map_1x_markings_to_2x,
@@ -76,6 +82,7 @@ from stix2elevator.vocab_mappings import (ATTACK_MOTIVATION_MAP,
                                           THREAT_ACTOR_LABEL_MAP,
                                           THREAT_ACTOR_SOPHISTICATION_MAP,
                                           TOOL_LABELS_MAP)
+
 
 if stix.__version__ >= "1.2.0.0":
     from stix.report import Report
@@ -353,158 +360,6 @@ def handle_free_text_lines(sdo_instance, free_text_lines):
             sdo_instance[convert_to_custom_property_name("free_text_lines")] = lines
 
 
-def add_string_property_to_description(sdo_instance, property_name, property_value, is_list=False):
-    if is_list:
-        sdo_instance["description"] += "\n\n" + property_name.upper() + ":\n"
-        property_values = []
-        for v in property_value:
-            property_values.append(text_type(v))
-        sdo_instance["description"] += ",\n".join(property_values)
-    else:
-        sdo_instance["description"] += "\n\n" + property_name.upper() + ":\n\t" + text_type(property_value)
-    warn("Appended %s to description of %s", 302, property_name, sdo_instance["id"])
-
-
-def add_string_property_as_custom_property(sdo_instance, property_name, property_value, is_list=False):
-    if is_list:
-        property_values = list()
-        for v in property_value:
-            property_values.append(text_type(v))
-        sdo_instance[convert_to_custom_property_name(property_name)] = ",".join(property_values)
-    else:
-        sdo_instance[convert_to_custom_property_name(property_name)] = text_type(property_value)
-    warn("Used custom property for %s of %s", 308, property_name, sdo_instance["id"])
-
-
-def handle_missing_string_property(sdo_instance, property_name, property_value, is_list=False):
-    if property_value:
-        if get_option_value("missing_policy") == "add-to-description":
-            add_string_property_to_description(sdo_instance, property_name, property_value, is_list)
-        elif get_option_value("missing_policy") == "use-custom-properties":
-            add_string_property_as_custom_property(sdo_instance, property_name, property_value, is_list)
-        else:
-            warn("Missing property %s of %s is ignored", 307, property_name, sdo_instance["id"])
-
-
-def add_confidence_property_to_description(sdo_instance, confidence, parent_property_name):
-    prefix = parent_property_name.upper() + " " if parent_property_name else ""
-    if confidence is not None:
-        sdo_instance["description"] += "\n\n" + prefix + "CONFIDENCE: "
-        if confidence.value is not None:
-            sdo_instance["description"] += text_type(confidence.value)
-        if confidence.description is not None:
-            sdo_instance["description"] += "\n\t" + prefix + "DESCRIPTION: " + text_type(confidence.description)
-        warn("Appended Confidence type content to description of %s", 304, sdo_instance["id"])
-
-
-def add_confidence_property_as_custom_property(sdo_instance, confidence, parent_property_name=None):
-    prefix = parent_property_name + "_" if parent_property_name else ""
-    if confidence.value is not None:
-        sdo_instance[convert_to_custom_property_name(prefix + "confidence")] = text_type(confidence.value)
-    if confidence.description is not None:
-        sdo_instance[convert_to_custom_property_name(prefix + "confidence_description")] = text_type(confidence.description)
-    warn("Used custom properties for Confidence type content of %s", 308, sdo_instance["id"])
-
-
-def handle_missing_confidence_property(sdo_instance, confidence, parent_property_name=None):
-    if confidence:
-        if get_option_value("missing_policy") == "add-to-description" and confidence:
-            add_confidence_property_to_description(sdo_instance, confidence, parent_property_name)
-        elif get_option_value("missing_policy") == "use-custom-properties":
-            add_confidence_property_as_custom_property(sdo_instance, confidence, parent_property_name)
-        else:
-            warn("Missing property confidence %s is ignored", 307, sdo_instance["id"])
-
-
-def add_statement_type_to_description(sdo_instance, statement, property_name):
-    sdo_instance["description"] += "\n\n" + property_name.upper() + ":"
-    has_value = False
-    if statement.value:
-        sdo_instance["description"] += text_type(statement.value)
-        has_value = True
-    if statement.descriptions:
-        descriptions = []
-        for d in statement.descriptions:
-            descriptions.append(text_type(d.value))
-        sdo_instance["description"] += (": " if has_value else "") + "\n\n\t".join(descriptions)
-    if statement.source is not None:
-        # FIXME: Handle source
-        info("Source in %s is not handled, yet.", 815, sdo_instance["id"])
-    if statement.confidence:
-        add_confidence_property_to_description(sdo_instance, statement.confidence, property_name)
-    warn("Appended Statement type content to description of %s", 305, sdo_instance["id"])
-
-
-def add_statement_type_as_custom_property(statement):
-    statement_json = {}
-    if statement.value:
-        statement_json["value"] = text_type(statement.value)
-    if statement.descriptions:
-        descriptions = []
-        for d in statement.descriptions:
-            descriptions.append(text_type(d.value))
-        statement_json["description"] = " ".join(descriptions)
-    if statement.source is not None:
-        # FIXME: Handle source
-        info("Source property in STIX 1.x statement is not handled, yet.", 815)
-    if statement.confidence:
-        add_confidence_property_as_custom_property(statement_json, statement.confidence)
-    return statement_json
-
-
-def statement_type_as_properties(sdo_instance, statement, property_name):
-    if statement.value:
-        sdo_instance[convert_to_custom_property_name(property_name)] = text_type(statement.value)
-    if statement.descriptions:
-        descriptions = []
-        for d in statement.descriptions:
-            descriptions.append(text_type(d.value))
-        sdo_instance[convert_to_custom_property_name(property_name) + "_description"] = " ".join(descriptions)
-    if statement.source is not None:
-        # FIXME: Handle source
-        info("Source property in STIX 1.x statement is not handled, yet.", 815)
-    if statement.confidence:
-        add_confidence_property_as_custom_property(sdo_instance, statement.confidence, property_name)
-
-
-def handle_missing_statement_properties(sdo_instance, statement, property_name):
-    if statement:
-        if get_option_value("missing_policy") == "add-to-description":
-            add_statement_type_to_description(sdo_instance, statement, property_name)
-        elif get_option_value("missing_policy") == "use-custom-properties":
-            statement_type_as_properties(sdo_instance, statement, property_name)
-            warn("Used custom properties for Statement type content of %s", 308, sdo_instance["id"])
-        else:
-            warn("Missing property %s of %s is ignored", 307, property_name, sdo_instance["id"])
-
-
-def handle_multiple_missing_statement_properties(sdo_instance, statements, property_name):
-    if statements:
-        if len(statements) == 1:
-            handle_missing_statement_properties(sdo_instance, statements[0], property_name)
-        else:
-            if get_option_value("missing_policy") == "add-to-description":
-                for s in statements:
-                    add_statement_type_to_description(sdo_instance, s, property_name)
-            elif get_option_value("missing_policy") == "use-custom-properties":
-                statements_json = []
-                for s in statements:
-                    statements_json.append(add_statement_type_as_custom_property(s))
-                sdo_instance[convert_to_custom_property_name(property_name + "s")] = statements_json
-            else:
-                warn("Missing property %s of %s is ignored", 307, property_name, sdo_instance["id"])
-
-
-def handle_missing_tool_property(sdo_instance, tool):
-    if tool.name:
-        if get_option_value("missing_policy") == "add-to-description":
-            sdo_instance["description"] += "\n\nTOOL SOURCE:"
-            sdo_instance["description"] += "\n\tname: " + text_type(tool.name)
-        warn("Appended Tool type content to description of %s", 306, sdo_instance["id"])
-    elif get_option_value("missing_policy") == "use-custom-properties":
-        sdo_instance[convert_to_custom_property_name("tool_source")] = text_type(tool.name)
-    else:
-        warn("Missing property name of %s is ignored", 307, sdo_instance["id"])
 
 
 # Sightings
